+++ spark-atlas-connector/src/main/scala/com/hortonworks/spark/atlas/sql/SparkExecutionPlanProcessor.scala
@@ -27,12 +27,13 @@ import org.apache.spark.sql.execution.datasources.v2.WriteToDataSourceV2Exec
 import org.apache.spark.sql.execution.streaming.sources.MicroBatchWriter
 import org.apache.spark.sql.hive.execution._
 import org.apache.spark.sql.sources.v2.writer.DataSourceWriter
-import org.apache.spark.sql.kafka010.atlas.KafkaHarvester
 import org.apache.spark.sql.kafka010.KafkaStreamWriter
 
 import com.hortonworks.spark.atlas.{AbstractEventProcessor, AtlasClient, AtlasClientConf}
 import com.hortonworks.spark.atlas.types.{external, metadata}
 import com.hortonworks.spark.atlas.utils.Logging
+import com.hortonworks.spark.atlas.sql.streaming.{HWCStreamingHarvester, KafkaHarvester}
+
 
 case class QueryDetail(qe: QueryExecution, executionId: Long,
   executionTime: Long, query: Option[String] = None)
