+++ spark-atlas-connector/src/main/scala/com/hortonworks/spark/atlas/sql/streaming/HWCStreamingHarvester.scala
@@ -15,12 +15,12 @@
  * limitations under the License.
  */
 
-package com.hortonworks.spark.atlas.sql
+package com.hortonworks.spark.atlas.sql.streaming
 
 import com.hortonworks.spark.atlas.sql.CommandsHarvester.HWCEntities
+import com.hortonworks.spark.atlas.sql.{Harvester, QueryDetail}
 import org.apache.atlas.model.instance.AtlasEntity
 import org.apache.spark.sql.execution.datasources.v2.WriteToDataSourceV2Exec
-import org.apache.spark.sql.kafka010.atlas.KafkaHarvester
 
 
 object HWCStreamingHarvester extends Harvester[WriteToDataSourceV2Exec] {
