+++ spark-atlas-connector/src/main/scala/com/hortonworks/spark/atlas/types/external.scala
@@ -22,6 +22,7 @@ import java.net.{URI, URISyntaxException}
 import java.util.Date
 
 import scala.collection.JavaConverters._
+
 import org.apache.atlas.AtlasConstants
 import org.apache.atlas.hbase.bridge.HBaseAtlasHook._
 import org.apache.atlas.model.instance.AtlasEntity
@@ -30,9 +31,11 @@ import org.apache.hadoop.fs.Path
 import org.apache.hadoop.hive.ql.session.SessionState
 import org.apache.spark.sql.catalyst.catalog.{CatalogDatabase, CatalogStorageFormat, CatalogTable}
 import org.apache.spark.sql.types.StructType
-import org.apache.spark.sql.kafka010.atlas.KafkaTopicInformation
+
 import com.hortonworks.spark.atlas.{AtlasClient, AtlasUtils}
 import com.hortonworks.spark.atlas.utils.SparkUtils
+import com.hortonworks.spark.atlas.sql.streaming.KafkaTopicInformation
+
 
 object external {
   // External metadata types used to link with external entities
